Sheriyans part 2/requirements.txt
# Core LLM + API clients
openai>=1.30.0
mistralai>=0.4.2

# LangChain ecosystem (RAG orchestration)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0

# Vector database
chromadb>=0.5.0

# Embeddings + transformers (for local fallback)
sentence-transformers>=2.7.0
transformers>=4.41.0
torch>=2.2.0

# File/document loaders
pypdf>=4.2.0
pdfplumber>=0.11.0
python-docx>=1.1.0
unstructured>=0.14.0

# Text processing
tiktoken>=0.7.0

# Backend (API)
fastapi>=0.111.0
uvicorn>=0.30.0

# Environment management
python-dotenv>=1.0.1

# Utilities
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0

# Async + HTTP
httpx>=0.27.0
aiofiles>=23.2.1

# Optional (streaming, better performance)
orjson>=3.10.0

# Logging & debugging
loguru>=0.7.2